requirements.txt
asgiref==3.8.1
django==5.2.3
django-filter==25.1
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
pip==24.0
pytz==2025.2
pyyaml==6.0.2
sqlparse==0.5.3
uritemplate==4.2.0
